+++ printlinter/ignored_processing.py
@@ -1,7 +1,7 @@
 """Processing ingoed lines functions."""
 
 # Local imports
-from .classes import IgnoreFile, IgnoreLine, IssueInfo
+from .classes import IgnoredBlock, IgnoreFile, IgnoreLine, IssueInfo
 
 FAMILIRY_ERR_CODE = [
     "PPL000",
@@ -21,6 +21,7 @@ def get_not_ignore_issue(
     issues: list[IssueInfo],
     ignore_lines: list[IgnoreLine],
     ignore_files: list[IgnoreFile],
+    ignored_blocks: list[IgnoredBlock],
     disabled_rules: list[str],
 ) -> list[IssueInfo]:
     """
@@ -30,6 +31,7 @@ def get_not_ignore_issue(
         issues: All issue found.
         ignore_lines: All ignored lines.
         ignore_files: All ignored files.
+        ignored_blocks: All ignored blocks of code.
         disabled_rules: All disabled rules.
 
     Returns:
@@ -77,6 +79,24 @@ def get_not_ignore_issue(
             from_file=issue.from_file,
         )
         if ignore_line_equivalence in ignore_lines:
+            continue
+
+        # Ignored blocks
+        # TODO: ignore a librairy rules for a block of code
+        issue_in_ignored_block = False
+        for ignored_block in ignored_blocks:
+            if (
+                ignored_block.from_file == issue.from_file
+                and ignored_block.line_from <= issue.num_line <= ignored_block.line_to
+                and (
+                    ignored_block.error_code == "ALL"
+                    or ignored_block.error_code == issue.issue.err_code
+                )
+            ):
+                issue_in_ignored_block = True
+                continue
+
+        if issue_in_ignored_block:
             continue
 
         result.append(issue)