+++ tests/core/test_ignored_processing.py
@@ -3,7 +3,7 @@ from printlinter import IssueEnum, IssueInfo, get_not_ignore_issue
 
 
 def test_get_not_ignore_issue__ignore_lines(issues_for_lines, ignored_lines):
-    not_ignored_line = get_not_ignore_issue(issues_for_lines, ignored_lines, [], [])
+    not_ignored_line = get_not_ignore_issue(issues_for_lines, ignored_lines, [], [], [])
 
     expected = [
         IssueInfo(
@@ -201,7 +201,7 @@ def test_get_not_ignore_issue__ignore_lines(issues_for_lines, ignored_lines):
 
 
 def test_get_not_ignore_issue__ignore_files(issues_for_files, ignored_files):
-    not_ignored_line = get_not_ignore_issue(issues_for_files, [], ignored_files, [])
+    not_ignored_line = get_not_ignore_issue(issues_for_files, [], ignored_files, [], [])
 
     expected = [
         IssueInfo(
@@ -347,6 +347,7 @@ def test_get_not_ignore_issue__ignore_files_and_lines(
         ignored_lines,
         ignored_files,
         [],
+        [],
     )
 
     expected = [
@@ -457,6 +458,269 @@ def test_get_not_ignore_issue__ignore_files_and_lines(
         IssueInfo(
             issue=IssueEnum.SYSSTDERRWRITELINESDETECT,
             num_line=6,
+            num_col=12,
+            line_as_str="sys.stderr.writelines(['toto', 'titi'])",
+            from_file="toto5.py",
+            ignore=False,
+        ),
+    ]
+
+    assert not_ignored_line == expected
+
+
+def test_get_not_ignore_issue__ignore_block(
+    issues_for_blocks,
+    ignored_blocks,
+):
+    not_ignored_line = get_not_ignore_issue(
+        issues_for_blocks,
+        [],
+        [],
+        ignored_blocks,
+        [],
+    )
+
+    expected = [
+        IssueInfo(
+            issue=IssueEnum.PRINTDETECT,
+            num_line=6,
+            num_col=12,
+            line_as_str="print('toto')",
+            from_file="toto2.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRINTDETECT,
+            num_line=6,
+            num_col=12,
+            line_as_str="print('toto')",
+            from_file="toto3.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRINTDETECT,
+            num_line=6,
+            num_col=12,
+            line_as_str="print('toto')",
+            from_file="toto4.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRINTDETECT,
+            num_line=6,
+            num_col=12,
+            line_as_str="print('toto')",
+            from_file="toto5.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRINTDETECT,
+            num_line=6,
+            num_col=12,
+            line_as_str="print('toto')",
+            from_file="toto6.py",
+            ignore=False,
+        ),
+        #
+        IssueInfo(
+            issue=IssueEnum.PRETTYPRINTDETECT,
+            num_line=7,
+            num_col=12,
+            line_as_str="pprint('toto')",
+            from_file="toto1.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRETTYPRINTDETECT,
+            num_line=7,
+            num_col=12,
+            line_as_str="pprint('toto')",
+            from_file="toto3.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRETTYPRINTDETECT,
+            num_line=7,
+            num_col=12,
+            line_as_str="pprint('toto')",
+            from_file="toto4.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRETTYPRINTDETECT,
+            num_line=7,
+            num_col=12,
+            line_as_str="pprint('toto')",
+            from_file="toto5.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.PRETTYPRINTDETECT,
+            num_line=7,
+            num_col=12,
+            line_as_str="pprint('toto')",
+            from_file="toto6.py",
+            ignore=False,
+        ),
+        #
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITEDETECT,
+            num_line=8,
+            num_col=12,
+            line_as_str="sys.stdout.write('toto')",
+            from_file="toto1.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITEDETECT,
+            num_line=8,
+            num_col=12,
+            line_as_str="sys.stdout.write('toto')",
+            from_file="toto2.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITEDETECT,
+            num_line=8,
+            num_col=12,
+            line_as_str="sys.stdout.write('toto')",
+            from_file="toto4.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITEDETECT,
+            num_line=8,
+            num_col=12,
+            line_as_str="sys.stdout.write('toto')",
+            from_file="toto5.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITEDETECT,
+            num_line=8,
+            num_col=12,
+            line_as_str="sys.stdout.write('toto')",
+            from_file="toto6.py",
+            ignore=False,
+        ),
+        #
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITEDETECT,
+            num_line=9,
+            num_col=12,
+            line_as_str="sys.stderr.write('toto')",
+            from_file="toto1.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITEDETECT,
+            num_line=9,
+            num_col=12,
+            line_as_str="sys.stderr.write('toto')",
+            from_file="toto2.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITEDETECT,
+            num_line=9,
+            num_col=12,
+            line_as_str="sys.stderr.write('toto')",
+            from_file="toto3.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITEDETECT,
+            num_line=9,
+            num_col=12,
+            line_as_str="sys.stderr.write('toto')",
+            from_file="toto5.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITEDETECT,
+            num_line=9,
+            num_col=12,
+            line_as_str="sys.stderr.write('toto')",
+            from_file="toto6.py",
+            ignore=False,
+        ),
+        #
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITELINESDETECT,
+            num_line=10,
+            num_col=12,
+            line_as_str="sys.stdout.writelines(['toto', 'titi'])",
+            from_file="toto1.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITELINESDETECT,
+            num_line=10,
+            num_col=12,
+            line_as_str="sys.stdout.writelines(['toto', 'titi'])",
+            from_file="toto2.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITELINESDETECT,
+            num_line=10,
+            num_col=12,
+            line_as_str="sys.stdout.writelines(['toto', 'titi'])",
+            from_file="toto3.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITELINESDETECT,
+            num_line=10,
+            num_col=12,
+            line_as_str="sys.stdout.writelines(['toto', 'titi'])",
+            from_file="toto4.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDOUTWRITELINESDETECT,
+            num_line=10,
+            num_col=12,
+            line_as_str="sys.stdout.writelines(['toto', 'titi'])",
+            from_file="toto6.py",
+            ignore=False,
+        ),
+        #
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITELINESDETECT,
+            num_line=11,
+            num_col=12,
+            line_as_str="sys.stderr.writelines(['toto', 'titi'])",
+            from_file="toto1.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITELINESDETECT,
+            num_line=11,
+            num_col=12,
+            line_as_str="sys.stderr.writelines(['toto', 'titi'])",
+            from_file="toto2.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITELINESDETECT,
+            num_line=11,
+            num_col=12,
+            line_as_str="sys.stderr.writelines(['toto', 'titi'])",
+            from_file="toto3.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITELINESDETECT,
+            num_line=11,
+            num_col=12,
+            line_as_str="sys.stderr.writelines(['toto', 'titi'])",
+            from_file="toto4.py",
+            ignore=False,
+        ),
+        IssueInfo(
+            issue=IssueEnum.SYSSTDERRWRITELINESDETECT,
+            num_line=11,
             num_col=12,
             line_as_str="sys.stderr.writelines(['toto', 'titi'])",
             from_file="toto5.py",