+++ tests/testing_files/ignored_block/ppl003.py
@@ -1,0 +1,12 @@
+# Standard imports
+from sys import stdout
+
+
+def toto():
+    return 1 + 2
+
+
+# <py-printlinter disable PPL003>
+stdout.write(toto())
+stdout.write(toto() + toto())
+# <py-printlinter enable PPL003>